revert: integration test module
