+++ test-aether-integration-module/conf/pip/primary-requirements.py3.txt
@@ -1,0 +1,22 @@
+#################################################################################################################
+#                                                                              					#
+#                             Primary requirements                             					#
+#                                                                              					#
+#                                 CHANGE ME !!!                                					#
+#                                                                              					#
+# Update me and then run 'docker-compose -f docker-compose-test.yml run integration-test pip_freeze'            #
+#                                                                              					#
+#################################################################################################################
+
+
+aether.client
+aether.mocker
+avro-python3
+coverage
+flake8
+kafka
+mock
+pytest
+pytest-runner
+requests
+spavro
