+++ test-aether-integration-module/conf/pip/requirements.py3.txt
@@ -1,0 +1,38 @@
+#########################################################################################
+#                                                                              		#
+#                              Frozen requirements                             		#
+#                                                                              		#
+#                               DO NOT CHANGE !!!                              		#
+#                                                                              		#
+# To update this first update 'primary-requirements.txt'                       		#
+# then run 'docker-compose -f docker-compose-test.yml run integration-test pip_freeze'	#
+#                                                                              		#
+#########################################################################################
+aether.client==0.0.0
+aether.mocker==0.0.0
+attrs==17.4.0
+avro-python3==1.8.2
+certifi==2018.1.18
+chardet==3.0.4
+configobj==5.0.6
+configparser==3.5.0
+coverage==4.5.1
+enum34==1.1.6
+flake8==3.5.0
+funcsigs==1.0.2
+idna==2.6
+kafka==1.3.5
+mccabe==0.6.1
+mock==2.0.0
+more-itertools==4.1.0
+pbr==4.0.1
+pluggy==0.6.0
+py==1.5.3
+pycodestyle==2.3.1
+pyflakes==1.6.0
+pytest==3.5.0
+pytest-runner==4.2
+requests==2.18.4
+six==1.11.0
+spavro==1.1.10
+urllib3==1.22
