+++ test-aether-integration-module/test/test-py2_schema_generation.py
@@ -1,0 +1,11 @@
+from . import *
+
+
+@py2
+def test_1_register_schemas(schema_registration):
+    assert(schema_registration)  # Try to use the test mode of the wizard
+
+
+@py2
+def test_2_check_schemas(existing_schemas):
+    assert(len(existing_schemas) > 0)
